@@ -29,24 +29,28 @@ setup(
         "Operating System :: Microsoft",
         "Programming Language :: Python",
         "Programming Language :: Python :: 3",
-        "Programming Language :: Python :: 3.5",
-        "Programming Language :: Python :: 3.6",
         "Programming Language :: Python :: 3.7",
         "Programming Language :: Python :: 3.8",
+        "Programming Language :: Python :: 3.9",
+        "Programming Language :: Python :: 3.10",
         "Programming Language :: Python :: Implementation :: PyPy",
         "Programming Language :: Python :: Implementation :: CPython",
     ],
     packages=find_packages(exclude=["tests"]),
     package_data={"aio_pika": ["py.typed"]},
-    install_requires=["aiormq>=3.2.3,<4", "yarl"],
-    python_requires=">=3.5, <4",
+    install_requires=[
+        "aiormq~=6.4.0",
+        "yarl",
+        'typing_extensions; python_version < "3.8.0"'
+    ],
+    python_requires=">3.6, <4",
     extras_require={
         "develop": [
-            "aiomisc~=10.1.6",
-            "async_generator",
+            "aiomisc~=16.0",
             "coverage!=4.3",
             "coveralls",
             "pylava",
+            "pyflakes<2.5",
             "pytest",
             "pytest-cov",
             "shortuuid",
