@@ -1,20 +1,22 @@
 import asyncio
-import aio_pika
 import ssl
 
+import aio_pika
+from aio_pika.abc import SSLOptions
 
-async def main(loop):
+
+async def main() -> None:
     connection = await aio_pika.connect_robust(
         host="127.0.0.1",
         login="",
         ssl=True,
-        ssl_options=dict(
-            ca_certs="cacert.pem",
+        ssl_options=SSLOptions(
+            cafile="cacert.pem",
             certfile="cert.pem",
             keyfile="key.pem",
-            cert_reqs=ssl.CERT_REQUIRED,
+            no_verify_ssl=ssl.CERT_REQUIRED,
         ),
-        loop=loop,
+        client_properties={"connection_name": "aio-pika external credentials"},
     )
 
     async with connection:
@@ -29,6 +31,4 @@ async def main(loop):
 
 
 if __name__ == "__main__":
-    loop = asyncio.get_event_loop()
-    loop.run_until_complete(main(loop))
-    loop.close()
+    asyncio.run(main())
