@@ -1,29 +1,31 @@
 import asyncio
+
 import aio_pika
+from aio_pika.abc import AbstractRobustConnection
 from aio_pika.pool import Pool
 
 
-async def main():
+async def main() -> None:
     loop = asyncio.get_event_loop()
 
-    async def get_connection():
+    async def get_connection() -> AbstractRobustConnection:
         return await aio_pika.connect_robust("amqp://guest:guest@localhost/")
 
-    connection_pool = Pool(get_connection, max_size=2, loop=loop)
+    connection_pool: Pool = Pool(get_connection, max_size=2, loop=loop)
 
     async def get_channel() -> aio_pika.Channel:
         async with connection_pool.acquire() as connection:
             return await connection.channel()
 
-    channel_pool = Pool(get_channel, max_size=10, loop=loop)
+    channel_pool: Pool = Pool(get_channel, max_size=10, loop=loop)
     queue_name = "pool_queue"
 
-    async def consume():
+    async def consume() -> None:
         async with channel_pool.acquire() as channel:  # type: aio_pika.Channel
             await channel.set_qos(10)
 
             queue = await channel.declare_queue(
-                queue_name, durable=False, auto_delete=False
+                queue_name, durable=False, auto_delete=False,
             )
 
             async with queue.iterator() as queue_iter:
@@ -31,7 +33,7 @@ async def main():
                     print(message)
                     await message.ack()
 
-    async def publish():
+    async def publish() -> None:
         async with channel_pool.acquire() as channel:  # type: aio_pika.Channel
             await channel.default_exchange.publish(
                 aio_pika.Message(("Channel: %r" % channel).encode()),
@@ -40,11 +42,9 @@ async def main():
 
     async with connection_pool, channel_pool:
         task = loop.create_task(consume())
-        await asyncio.wait([publish() for _ in range(10000)])
+        await asyncio.wait([publish() for _ in range(50)])
         await task
 
 
 if __name__ == "__main__":
-    loop = asyncio.get_event_loop()
-    loop.run_until_complete(main())
-    loop.close()
+    asyncio.run(main())
