@@ -1,13 +1,14 @@
 import asyncio
+
 from aio_pika import connect_robust
 from aio_pika.patterns import RPC
 
 
-async def multiply(*, x, y):
+async def multiply(*, x: int, y: int) -> int:
     return x * y
 
 
-async def main():
+async def main() -> None:
     connection = await connect_robust(
         "amqp://guest:guest@127.0.0.1/",
         client_properties={"connection_name": "callee"},
@@ -19,15 +20,12 @@ async def main():
     rpc = await RPC.create(channel)
     await rpc.register("multiply", multiply, auto_delete=True)
 
-    return connection
+    try:
+        await asyncio.Future()
+    finally:
+        await connection.close()
 
 
 if __name__ == "__main__":
-    loop = asyncio.get_event_loop()
-    connection = loop.run_until_complete(main())
+    asyncio.run(main())
 
-    try:
-        loop.run_forever()
-    finally:
-        loop.run_until_complete(connection.close())
-        loop.shutdown_asyncgens()
