@@ -1,10 +1,11 @@
 import asyncio
+
 import aio_pika
 
 
-async def main(loop):
+async def main() -> None:
     connection = await aio_pika.connect_robust(
-        "amqp://guest:guest@127.0.0.1/", loop=loop
+        "amqp://guest:guest@127.0.0.1/",
     )
 
     async with connection:
@@ -13,12 +14,10 @@ async def main(loop):
         channel = await connection.channel()
 
         await channel.default_exchange.publish(
-            aio_pika.Message(body="Hello {}".format(routing_key).encode()),
+            aio_pika.Message(body=f"Hello {routing_key}".encode()),
             routing_key=routing_key,
         )
 
 
 if __name__ == "__main__":
-    loop = asyncio.get_event_loop()
-    loop.run_until_complete(main(loop))
-    loop.close()
+    asyncio.run(main())
