@@ -1,9 +1,11 @@
 import asyncio
+
 from aio_pika import connect_robust
-from aio_pika.patterns import Master, RejectMessage, NackMessage
+from aio_pika.abc import AbstractConnection
+from aio_pika.patterns import Master, NackMessage, RejectMessage
 
 
-async def worker(*, task_id):
+async def worker(*, task_id: int) -> None:
     # If you want to reject message or send
     # nack you might raise special exception
 
@@ -16,22 +18,23 @@ async def worker(*, task_id):
     print(task_id)
 
 
-async def main():
-    connection = await connect_robust("amqp://guest:guest@127.0.0.1/")
+async def main() -> None:
+    connection = await connect_robust(
+        "amqp://guest:guest@127.0.0.1/?name=aio-pika%20worker",
+    )
 
     # Creating channel
     channel = await connection.channel()
 
+    # Initializing Master with channel
     master = Master(channel)
     await master.create_worker("my_task_name", worker, auto_delete=True)
 
-    return connection
+    try:
+        await asyncio.Future()
+    finally:
+        await connection.close()
 
 
 if __name__ == "__main__":
-    loop = asyncio.get_event_loop()
-    connection = loop.run_until_complete(main())
-    try:
-        loop.run_forever()
-    finally:
-        loop.run_until_complete(connection.close())
+    asyncio.run(main())
