@@ -1,40 +1,38 @@
 import asyncio
-from aio_pika import connect, IncomingMessage
+
+from aio_pika import connect
+from aio_pika.abc import AbstractIncomingMessage
 
 
-async def on_message(message: IncomingMessage):
+async def on_message(message: AbstractIncomingMessage) -> None:
     """
     on_message doesn't necessarily have to be defined as async.
     Here it is to show that it's possible.
     """
     print(" [x] Received message %r" % message)
     print("Message body is: %r" % message.body)
+
     print("Before sleep!")
     await asyncio.sleep(5)  # Represents async I/O operations
     print("After sleep!")
 
 
-async def main(loop):
+async def main() -> None:
     # Perform connection
-    connection = await connect(
-        "amqp://guest:guest@localhost/", loop=loop
-    )
+    connection = await connect("amqp://guest:guest@localhost/")
+    async with connection:
+        # Creating a channel
+        channel = await connection.channel()
 
-    # Creating a channel
-    channel = await connection.channel()
+        # Declaring queue
+        queue = await channel.declare_queue("hello")
 
-    # Declaring queue
-    queue = await channel.declare_queue("hello")
+        # Start listening the queue with name 'hello'
+        await queue.consume(on_message, no_ack=True)
 
-    # Start listening the queue with name 'hello'
-    await queue.consume(on_message, no_ack=True)
+        print(" [*] Waiting for messages. To exit press CTRL+C")
+        await asyncio.Future()
 
 
 if __name__ == "__main__":
-    loop = asyncio.get_event_loop()
-    loop.create_task(main(loop))
-
-    # we enter a never-ending loop that waits for data and
-    # runs callbacks whenever necessary.
-    print(" [*] Waiting for messages. To exit press CTRL+C")
-    loop.run_forever()
+    asyncio.run(main())
