@@ -1,38 +1,36 @@
 import asyncio
-from aio_pika import connect, IncomingMessage
 
-loop = asyncio.get_event_loop()
+from aio_pika import connect
+from aio_pika.abc import AbstractIncomingMessage
 
 
-def on_message(message: IncomingMessage):
+async def on_message(message: AbstractIncomingMessage) -> None:
     async with message.process():
-        print(" [x] Received message %r" % message)
-        print("     Message body is: %r" % message.body)
+        print(f" [x] Received message {message!r}")
+        print(f"     Message body is: {message.body!r}")
 
 
-async def main():
+async def main() -> None:
     # Perform connection
-    connection = await connect("amqp://guest:guest@localhost/", loop=loop)
+    connection = await connect("amqp://guest:guest@localhost/")
 
-    # Creating a channel
-    channel = await connection.channel()
-    await channel.set_qos(prefetch_count=1)
+    async with connection:
+        # Creating a channel
+        channel = await connection.channel()
+        await channel.set_qos(prefetch_count=1)
 
-    # Declaring queue
-    queue = await channel.declare_queue(
-        "task_queue",
-        durable=True
-    )
+        # Declaring queue
+        queue = await channel.declare_queue(
+            "task_queue",
+            durable=True,
+        )
 
-    # Start listening the queue with name 'task_queue'
-    await queue.consume(on_message)
+        # Start listening the queue with name 'task_queue'
+        await queue.consume(on_message)
+
+        print(" [*] Waiting for messages. To exit press CTRL+C")
+        await asyncio.Future()
 
 
 if __name__ == "__main__":
-    loop = asyncio.get_event_loop()
-    loop.create_task(main())
-
-    # we enter a never-ending loop that waits for data and runs
-    # callbacks whenever necessary.
-    print(" [*] Waiting for messages. To exit press CTRL+C")
-    loop.run_forever()
+    asyncio.run(main())
