@@ -1,37 +1,35 @@
-import sys
 import asyncio
-from aio_pika import connect, Message, DeliveryMode, ExchangeType
+import sys
+
+from aio_pika import DeliveryMode, ExchangeType, Message, connect
 
 
-async def main(loop):
+async def main() -> None:
     # Perform connection
-    connection = await connect(
-        "amqp://guest:guest@localhost/", loop=loop
-    )
+    connection = await connect("amqp://guest:guest@localhost/")
 
-    # Creating a channel
-    channel = await connection.channel()
+    async with connection:
+        # Creating a channel
+        channel = await connection.channel()
 
-    logs_exchange = await channel.declare_exchange(
-        "logs", ExchangeType.FANOUT
-    )
+        logs_exchange = await channel.declare_exchange(
+            "logs", ExchangeType.FANOUT,
+        )
 
-    message_body = b" ".join(
-        arg.encode() for arg in sys.argv[1:]) or b"Hello World!"
+        message_body = b" ".join(
+            arg.encode() for arg in sys.argv[1:]
+        ) or b"Hello World!"
 
-    message = Message(
-        message_body,
-        delivery_mode=DeliveryMode.PERSISTENT
-    )
+        message = Message(
+            message_body,
+            delivery_mode=DeliveryMode.PERSISTENT,
+        )
 
-    # Sending the message
-    await logs_exchange.publish(message, routing_key="info")
+        # Sending the message
+        await logs_exchange.publish(message, routing_key="info")
 
-    print(" [x] Sent %r" % message)
-
-    await connection.close()
+        print(f" [x] Sent {message!r}")
 
 
 if __name__ == "__main__":
-    loop = asyncio.get_event_loop()
-    loop.run_until_complete(main(loop))
+    asyncio.run(main())
