@@ -1,43 +1,39 @@
 import asyncio
-from aio_pika import connect, IncomingMessage, ExchangeType
 
-loop = asyncio.get_event_loop()
+from aio_pika import ExchangeType, connect
+from aio_pika.abc import AbstractIncomingMessage
 
 
-async def on_message(message: IncomingMessage):
+async def on_message(message: AbstractIncomingMessage) -> None:
     async with message.process():
-        print("[x] %r" % message.body)
+        print(f"[x] {message.body!r}")
 
 
-async def main():
+async def main() -> None:
     # Perform connection
-    connection = await connect(
-        "amqp://guest:guest@localhost/", loop=loop
-    )
+    connection = await connect("amqp://guest:guest@localhost/")
 
-    # Creating a channel
-    channel = await connection.channel()
-    await channel.set_qos(prefetch_count=1)
+    async with connection:
+        # Creating a channel
+        channel = await connection.channel()
+        await channel.set_qos(prefetch_count=1)
 
-    logs_exchange = await channel.declare_exchange(
-        "logs", ExchangeType.FANOUT
-    )
+        logs_exchange = await channel.declare_exchange(
+            "logs", ExchangeType.FANOUT,
+        )
 
-    # Declaring queue
-    queue = await channel.declare_queue(exclusive=True)
+        # Declaring queue
+        queue = await channel.declare_queue(exclusive=True)
 
-    # Binding the queue to the exchange
-    await queue.bind(logs_exchange)
+        # Binding the queue to the exchange
+        await queue.bind(logs_exchange)
 
-    # Start listening the queue with name 'task_queue'
-    await queue.consume(on_message)
+        # Start listening the queue
+        await queue.consume(on_message)
+
+        print(" [*] Waiting for logs. To exit press CTRL+C")
+        await asyncio.Future()
 
 
 if __name__ == "__main__":
-    loop = asyncio.get_event_loop()
-    loop.create_task(main())
-
-    # we enter a never-ending loop that waits for data
-    # and runs callbacks whenever necessary.
-    print(" [*] Waiting for logs. To exit press CTRL+C")
-    loop.run_forever()
+    asyncio.run(main())
