@@ -1,39 +1,39 @@
-import sys
 import asyncio
-from aio_pika import connect, Message, DeliveryMode, ExchangeType
+import sys
+
+from aio_pika import DeliveryMode, ExchangeType, Message, connect
 
 
-async def main(loop):
+async def main() -> None:
     # Perform connection
     connection = await connect(
-        "amqp://guest:guest@localhost/", loop=loop
+        "amqp://guest:guest@localhost/",
     )
 
-    # Creating a channel
-    channel = await connection.channel()
+    async with connection:
+        # Creating a channel
+        channel = await connection.channel()
 
-    topic_logs_exchange = await channel.declare_exchange(
-        "topic_logs", ExchangeType.TOPIC
-    )
+        topic_logs_exchange = await channel.declare_exchange(
+            "topic_logs", ExchangeType.TOPIC,
+        )
 
-    routing_key = sys.argv[1] if len(sys.argv) > 2 else "anonymous.info"
+        routing_key = sys.argv[1] if len(sys.argv) > 2 else "anonymous.info"
 
-    message_body = b" ".join(
-        arg.encode() for arg in sys.argv[2:]) or b"Hello World!"
+        message_body = b" ".join(
+            arg.encode() for arg in sys.argv[2:]
+        ) or b"Hello World!"
 
-    message = Message(
-        message_body,
-        delivery_mode=DeliveryMode.PERSISTENT
-    )
+        message = Message(
+            message_body,
+            delivery_mode=DeliveryMode.PERSISTENT,
+        )
 
-    # Sending the message
-    await topic_logs_exchange.publish(message, routing_key=routing_key)
+        # Sending the message
+        await topic_logs_exchange.publish(message, routing_key=routing_key)
 
-    print(" [x] Sent %r" % message)
-
-    await connection.close()
+        print(f" [x] Sent {message!r}")
 
 
 if __name__ == "__main__":
-    loop = asyncio.get_event_loop()
-    loop.run_until_complete(main(loop))
+    asyncio.run(main())
