@@ -1,34 +1,37 @@
-from logging import getLogger
-from typing import Optional
+import asyncio
+from typing import Any, Dict, Union
 
 import aiormq
+from pamqp.common import Arguments
 
-from .channel import Channel
+from .abc import (
+    AbstractExchange, AbstractRobustExchange, ExchangeParamType, TimeoutType,
+)
 from .exchange import Exchange, ExchangeType
+from .log import get_logger
 
 
-log = getLogger(__name__)
+log = get_logger(__name__)
 
 
-class RobustExchange(Exchange):
+class RobustExchange(Exchange, AbstractRobustExchange):
     """ Exchange abstraction """
+
+    _bindings: Dict[Union[AbstractExchange, str], Dict[str, Any]]
 
     def __init__(
         self,
-        connection,
-        channel: aiormq.Channel,
+        channel: aiormq.abc.AbstractChannel,
         name: str,
-        type: ExchangeType = ExchangeType.DIRECT,
+        type: Union[ExchangeType, str] = ExchangeType.DIRECT,
         *,
-        auto_delete: Optional[bool],
-        durable: Optional[bool],
-        internal: Optional[bool],
-        passive: Optional[bool],
-        arguments: dict = None
+        auto_delete: bool = False,
+        durable: bool = False,
+        internal: bool = False,
+        passive: bool = False,
+        arguments: Arguments = None
     ):
-
         super().__init__(
-            connection=connection,
             channel=channel,
             name=name,
             type=type,
@@ -39,28 +42,35 @@ class RobustExchange(Exchange):
             arguments=arguments,
         )
 
-        self._bindings = dict()
+        self._bindings = {}
+        self.__restore_lock = asyncio.Lock()
 
-    async def restore(self, channel: Channel):
-        self._channel = channel._channel
+    async def restore(self, channel: aiormq.abc.AbstractChannel) -> None:
+        async with self.__restore_lock:
+            try:
+                self.channel = channel
 
-        if self.name == "":
-            return
+                # special case for default exchange
+                if self.name == "":
+                    return
 
-        await self.declare()
+                await self.declare()
 
-        for exchange, kwargs in self._bindings.items():
-            await self.bind(exchange, **kwargs)
+                for exchange, kwargs in tuple(self._bindings.items()):
+                    await self.bind(exchange, **kwargs)
+            except Exception:
+                del self.channel
+                raise
 
     async def bind(
         self,
-        exchange,
+        exchange: ExchangeParamType,
         routing_key: str = "",
         *,
-        arguments=None,
-        timeout: int = None,
+        arguments: Arguments = None,
+        timeout: TimeoutType = None,
         robust: bool = True
-    ):
+    ) -> aiormq.spec.Exchange.BindOk:
         result = await super().bind(
             exchange,
             routing_key=routing_key,
@@ -70,19 +80,19 @@ class RobustExchange(Exchange):
 
         if robust:
             self._bindings[exchange] = dict(
-                routing_key=routing_key, arguments=arguments,
+                routing_key=routing_key,
+                arguments=arguments,
             )
 
         return result
 
     async def unbind(
         self,
-        exchange,
+        exchange: ExchangeParamType,
         routing_key: str = "",
-        arguments: dict = None,
-        timeout: int = None,
-    ):
-
+        arguments: Arguments = None,
+        timeout: TimeoutType = None,
+    ) -> aiormq.spec.Exchange.UnbindOk:
         result = await super().unbind(
             exchange, routing_key, arguments=arguments, timeout=timeout,
         )
