@@ -17,9 +17,9 @@
 # add these directories to sys.path here. If the directory is relative to the
 # documentation root, use os.path.abspath to make it absolute, like shown here.
 #
+import datetime
 import os
 import sys
-
 from importlib.machinery import SourceFileLoader
 
 
@@ -70,7 +70,7 @@ master_doc = "index"
 
 # General information about the project.
 project = "aio-pika"
-copyright = "2018, Dmitry Orlov"
+copyright = "{}, Dmitry Orlov".format(datetime.datetime.now().year)
 author = "Dmitry Orlov"
 
 # The version info for the project you're documenting, acts as replacement for
@@ -106,7 +106,7 @@ todo_include_todos = False
 # The theme to use for HTML and HTML Help pages.  See the documentation for
 # a list of builtin themes.
 #
-html_theme = "alabaster"
+html_theme = "furo"
 
 # Theme options are theme-specific and customize the look and feel of a theme
 # further.  For a list of options available for each theme, see the
@@ -114,16 +114,36 @@ html_theme = "alabaster"
 #
 # html_theme_options = {}
 html_theme_options = {
-    "codecov_button": True,
-    "description": "Wrapper for the aiormq for asyncio and humans",
-    "github_banner": True,
-    "github_button": True,
-    "github_repo": "aio-pika",
-    "github_type": "star",
-    "github_user": "mosquito",
-    "travis_button": True,
+    "sidebar_hide_name": True,
+    "source_repository": "https://github.com/mosquito/aio-pika/",
+    "source_branch": "master",
+    "source_directory": "docs/source",
+    "footer_icons": [
+        {
+            "name": "GitHub",
+            "url": "https://github.com/mosquito/aio-pika",
+            "html": (
+                """<svg stroke="currentColor" fill="currentColor" """
+                """stroke-width="0" viewBox="0 0 16 16">"""
+                """<path fill-rule="evenodd" d="M8 0C3.58 0 0 3.58 0 8c0 """
+                """3.54 2.29 6.53 5.47 7.59.4.07.55-.17.55-.38 0-.19-.01-"""
+                """.82-.01-1.49-2.01.37-2.53-.49-2.69-.94-.09-.23-.48-.94"""
+                """-.82-1.13-.28-.15-.68-.52-.01-.53.63-.01 1.08.58 1.23."""
+                """82.72 1.21 1.87.87 2.33.66.07-.52.28-.87.51-1.07-1.78-"""
+                """.2-3.64-.89-3.64-3.95 0-.87.31-1.59.82-2.15-.08-.2-.36"""
+                """-1.02.08-2.12 0 0 .67-.21 2.2.82.64-.18 1.32-.27 2-.27"""
+                """.68 0 1.36.09 2 .27 1.53-1.04 2.2-.82 2.2-.82.44 1.1.1"""
+                """6 1.92.08 2.12.51.56.82 1.27.82 2.15 0 3.07-1.87 3.75-"""
+                """3.65 3.95.29.25.54.73.54 1.48 0 1.07-.01 1.93-.01 2.2 """
+                """0 .21.15.46.55.38A8.013 8.013 0 0 0 16 8c0-4.42-3.58-8"""
+                """-8-8z"></path></svg>"""
+            ),
+            "class": "",
+        },
+    ],
 }
 
+html_title = "Wrapper for the aiormq for asyncio and humans"
 
 # Add any paths that contain custom static files (such as style sheets) here,
 # relative to this directory. They are copied after the builtin static files,
@@ -217,4 +237,4 @@ epub_exclude_files = ["search.html"]
 html_logo = "_static/logo2x.png"
 html_favicon = "_static/icon.png"
 
-html_sidebars = {"**": ["about.html", "navigation.html", "searchbox.html"]}
+# html_sidebars = {"**": ["about.html", "navigation.html", "searchbox.html"]}
