@@ -1,25 +1,23 @@
 import asyncio
+
 import tornado.ioloop
 import tornado.web
 
-from aio_pika import connect_robust, Message
-
-tornado.ioloop.IOLoop.configure("tornado.platform.asyncio.AsyncIOLoop")
-io_loop = tornado.ioloop.IOLoop.current()
-asyncio.set_event_loop(io_loop.asyncio_loop)
+from aio_pika import Message, connect_robust
 
 
-QUEUE = asyncio.Queue()
+class Base:
+    QUEUE: asyncio.Queue
 
 
-class SubscriberHandler(tornado.web.RequestHandler):
-    async def get(self):
-        message = await QUEUE.get()
-        self.finish(message.body)
+class SubscriberHandler(tornado.web.RequestHandler, Base):
+    async def get(self) -> None:
+        message = await self.QUEUE.get()
+        await self.finish(message.body)
 
 
 class PublisherHandler(tornado.web.RequestHandler):
-    async def post(self):
+    async def post(self) -> None:
         connection = self.application.settings["amqp_connection"]
         channel = await connection.channel()
 
@@ -30,15 +28,17 @@ class PublisherHandler(tornado.web.RequestHandler):
         finally:
             await channel.close()
 
-        self.finish("OK")
+        await self.finish("OK")
 
 
-async def make_app():
+async def make_app() -> tornado.web.Application:
     amqp_connection = await connect_robust()
 
     channel = await amqp_connection.channel()
     queue = await channel.declare_queue("test", auto_delete=True)
-    await queue.consume(QUEUE.put, no_ack=True)
+    Base.QUEUE = asyncio.Queue()
+
+    await queue.consume(Base.QUEUE.put, no_ack=True)
 
     return tornado.web.Application(
         [(r"/publish", PublisherHandler), (r"/subscribe", SubscriberHandler)],
@@ -46,8 +46,12 @@ async def make_app():
     )
 
 
-if __name__ == "__main__":
-    app = io_loop.asyncio_loop.run_until_complete(make_app())
+async def main() -> None:
+    app = await make_app()
     app.listen(8888)
+    await asyncio.Future()
 
-    tornado.ioloop.IOLoop.current().start()
+
+if __name__ == "__main__":
+    asyncio.run(main())
+
